Add router and hook up search by id
