+++ client/src/index.js
@@ -1,45 +1,11 @@
 import React from "react";
 import ReactDOM from "react-dom";
-
-import thunkMiddleware from "redux-thunk";
-import { createLogger } from "redux-logger";
-import { createStore, applyMiddleware } from "redux";
-import { Provider } from "react-redux";
-
-import rootReducer from "./reducers";
-import App from "./App";
-import Search from "./components/search"
-import Auth from "./components/authentication/auth";
 import * as serviceWorker from "./serviceWorker";
-import {createMuiTheme} from "@material-ui/core/styles";
-import { ThemeProvider } from '@material-ui/styles';
 
-const loggerMiddleware = createLogger();
-
-const store = createStore(
-    rootReducer,
-    applyMiddleware(
-        thunkMiddleware, // lets us dispatch() functions
-        loggerMiddleware // neat middleware that logs actions
-    )
-);
-
-const theme = createMuiTheme({
-    palette: {
-        primary: {
-            main: '#00675b',
-        },
-    },
-});
+import App from "./App"
 
 ReactDOM.render(
-    <Provider store={store}>
-        <ThemeProvider theme={theme}>
-            <Auth>
-                <Search />
-            </Auth>
-        </ThemeProvider>
-    </Provider>,
+    <App/>,
     document.getElementById("root")
 );
